pyproject.toml
[project]
name = "jarvis-v2"
version = "0.1.0"
description = "Intelligent AI Assistant with multi-agent orchestration"
requires-python = ">=3.12"
dependencies = [
    # Core
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",

    # Database
    "supabase>=2.0.0",
    "redis>=5.0.0",
    "pgvector>=0.3.0",

    # Google APIs
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-genai>=1.0.0",

    # External APIs
    "httpx>=0.27.0",
    "openai>=1.0.0",

    # Telegram
    "python-telegram-bot>=21.0",

    # Web Scraping
    "crawl4ai>=0.3.0",

    # Utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jarvis"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
